requirements.txt
pdfminer.six
PyMuPDF
pillow
tabula-py